Removed deactivation alterations of Jetpack
Not allowing Jetpack to go through its own deactivation sequence is not recommended. There is cleanup that Jetpack and WordPress.com need to do, including turning off services on the WordPress.com side that will repeatedly hit your server.

I did notice in Jetpack the Heartbeat removal is in a strange location and have corrected that already to be part of the deactivation sequence. You should see it in an upcoming release, thus also rendering your Heartbeat disconnect irrelevant.